Created StorageFactory contract
A new contract StorageFactory was created in the same directory and the previous contract SimpleStorage was imported.
A new function was created to deploy the SimpleStorage contract in the new StorageFactory contract.

Problem:
The compiler tab prompted an error, "Unused local variable. SimpleStorage simpleStorage = new SimpleStorage();"
The deploy tab prompted, "Currently you have no contract instances to interact with."